requirements.txt
alabaster==0.7.12
asgiref==3.5.2
Babel==2.10.3
bootstrap-py==1.0.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==38.0.1
cycler==0.11.0
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-phone-field==1.8.1
docutils==0.19
fonttools==4.37.2
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==3.3
imagesize==1.4.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mysqlclient==2.1.1
numpy==1.23.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.4
pandas-datareader==0.10.0
pguard==0.4.0
Pillow==9.2.0
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
xlwt==1.3.0
